Separate auto-thread/link-check
